requirements.txt
opencv_python==4.5.2.54
tqdm==4.61.1
numpy==1.19.5
Pillow==8.2.0
tensorboard==2.5.0
imutils==0.5.4